Updated to only update remote files if requested
FossilOrigin-Name: 9e821bdcac15994e4bf74472ee35081d430eb983

+++ build/autogen.sh
@@ -1,17 +1,19 @@
 #! /bin/bash
 
-(
-	mkdir aclocal >/dev/null 2>/dev/null
-	cd aclocal || exit 1
+if [ "$1" = '--update' ]; then
+	(
+		mkdir aclocal >/dev/null 2>/dev/null
+		cd aclocal || exit 1
 
-	for file in shobj.m4 tcl.m4; do
-		rm -f "${file}"
+		for file in shobj.m4 tcl.m4; do
+			rm -f "${file}"
 
-		wget -O "${file}.new" "http://rkeene.org/devel/autoconf/${file}" || continue
+			wget -O "${file}.new" "http://rkeene.org/devel/autoconf/${file}" || continue
 
-		mv "${file}.new" "${file}"
-	done
-)
+			mv "${file}.new" "${file}"
+		done
+	)
+fi
 
 cat aclocal/*.m4 > aclocal.m4 || exit 1
 autoconf || exit 1
